# api/docs/API_SPECIFICATION.yaml
openapi: 3.0.3
info:
  title: AstraPBX API
  description: |
    AstraPBX is a multi-tenant cloud PBX platform built on Asterisk. This API provides complete control over telephony infrastructure — from provisioning SIP trunks and DIDs to managing call queues, users, greetings, and hold music.

    ## Authentication

    AstraPBX supports two authentication methods. Set your credentials once using the **Authorize** button above — they persist across all requests in this session.

    ### Option 1: API Key (Recommended)
    Pass your organization's API key and secret as headers. No login step required.
    ```
    X-API-Key: org_xxxxxxxxxxxx
    X-API-Secret: your-api-secret
    ```

    ### Option 2: Bearer Token (JWT)
    Authenticate via `POST /auth/login` with your API key and secret to receive a JWT token. Pass it as a Bearer token in subsequent requests.
    ```
    Authorization: Bearer eyJhbGciOi...
    ```

    ### Admin Endpoints
    Endpoints under `/admin/*` require admin credentials and are used only for initial organization setup. Do not use admin tokens for regular API operations.

    ## Auto-Deployment

    Most write operations (creating users, updating queues, changing routing) automatically deploy configuration changes to Asterisk and reload. No manual deployment step is needed unless you want to force a full redeploy via `/config/deploy`.

    ## Quick Start

    1. **Authenticate** — Use the Authorize button with your API Key + Secret
...

  version: 1.0.0
...
servers:
...
security:
  - ApiKeyAuth: []
    ApiSecretAuth: []
  - BearerAuth: []
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Organization API key (e.g., `org_<32-char-hex>`)
    ApiSecretAuth:
      type: apiKey
      in: header
      name: X-API-Secret
      description: Organization API secret
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from `POST /auth/login`
  schemas:
...
    Queue:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "q1u2e3u4-5678-90ab-cdef-1234567890ab"
        org_id:
          type: string
          format: uuid
          example: "2c662bff-8f80-483a-8235-74fd48965a9c"
        name:
          type: string
          example: "Support Queue"
        number:
          type: string
          example: "5000"
        strategy:
          type: string
          enum: [ringall, leastrecent, fewestcalls, random, rrmemory, linear]
          example: "ringall"
        timeout:
          type: integer
          example: 30
        max_wait_time:
          type: integer
          example: 300
        music_on_hold:
          type: string
          example: "default"
        ring_sound:
          type: string
          example: "ring"
        recording_enabled:
          type: boolean
          example: true
        retry:
          type: integer
          example: 5
        status:
          type: string
          enum: [active, inactive]
          example: "active"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    QueueMember:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "m1e2m3b4-5678-90ab-cdef-1234567890ab"
        queue_id:
          type: string
          format: uuid
          example: "q1u2e3u4-5678-90ab-cdef-1234567890ab"
        user_id:
          type: string
          format: uuid
          example: "u1s2e3r4-5678-90ab-cdef-1234567890ab"
        penalty:
          type: integer
          example: 0
        paused:
          type: boolean
          example: false
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /queues:
    get:
      security:
        - BearerAuth: []
      tags: [Queues]
      summary: List all queues
      description: |
        Retrieve all call queues for the organization

        **Example curl:**
        ```bash
        curl -X GET ${YOUR_SERVER_URL}/api/v1/queues \
          -H "Authorization: Bearer YOUR_ORG_TOKEN"
        ```
      responses:
        '200':
          description: List of queues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Queue'
    post:
      security:
        - BearerAuth: []
      tags: [Queues]
      summary: Create new queue
      description: |
        Create a new call queue for distributing calls to agents.

        **Example curl:**
        ```bash
        curl -X POST ${YOUR_SERVER_URL}/api/v1/queues \
          -H "Authorization: Bearer YOUR_ORG_TOKEN" \
          -H "Content-Type: application/json" \
          -d '{
            "name": "Support Queue",
            "number": "5000",
            "strategy": "ringall",
            "timeout": 30,
            "music_on_hold": "default",
            "recording_enabled": true
          }'
        ```
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, number]
              properties:
                name:
                  type: string
                  example: "Support Queue"
                number:
                  type: string
                  example: "5000"
                strategy:
                  type: string
                  enum: [ringall, leastrecent, fewestcalls, random, rrmemory, linear]
                  default: ringall
                  example: "ringall"
                timeout:
                  type: integer
                  default: 15
                  example: 30
                max_wait_time:
                  type: integer
                  default: 300
                  example: 300
                music_on_hold:
                  type: string
                  default: default
                  example: "default"
                ring_sound:
                  type: string
                  default: ring
                  example: "ring"
                recording_enabled:
                  type: boolean
                  default: false
                  example: true
            example:
              name: "Support Queue"
              number: "5000"
              strategy: "ringall"
              timeout: 30
              music_on_hold: "default"
              recording_enabled: true
      responses:
        '201':
          description: Queue created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Queue'
  /queues/{queueId}:
    get:
      security:
        - BearerAuth: []
      tags: [Queues]
      summary: Get queue details
      description: |
        Retrieve details of a specific queue

        **Example curl:**
        ```bash
        curl -X GET ${YOUR_SERVER_URL}/api/v1/queues/QUEUE_UUID \
          -H "Authorization: Bearer YOUR_ORG_TOKEN"
        ```
      parameters:
        - name: queueId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Queue details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Queue'
    put:
      security:
        - BearerAuth: []
      tags: [Queues]
      summary: Update queue
      description: |
        Update an existing queue

        **Example curl:**
        ```bash
        curl -X PUT ${YOUR_SERVER_URL}/api/v1/queues/QUEUE_UUID \
          -H "Authorization: Bearer YOUR_ORG_TOKEN" \
          -H "Content-Type: application/json" \
          -d '{
            "name": "Updated Support Queue",
            "number": "5000",
            "strategy": "fewestcalls",
            "timeout": 45
          }'
        ```
      parameters:
        - name: queueId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Queue'
      responses:
        '200':
          description: Queue updated
    delete:
      security:
        - BearerAuth: []
      tags: [Queues]
      summary: Delete queue
      description: |
        Delete a queue

        **Example curl:**
        ```bash
        curl -X DELETE ${YOUR_SERVER_URL}/api/v1/queues/QUEUE_UUID \
          -H "Authorization: Bearer YOUR_ORG_TOKEN"
        ```
      parameters:
        - name: queueId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Queue deleted
  /queues/{queueId}/members:
    post:
      security:
        - BearerAuth: []
      tags: [Queues]
      summary: Add member to queue
      description: |
        Add a user as a member to the specified queue. This endpoint automatically
        triggers configuration deployment to ensure the new queue member is
        immediately available in the Asterisk system.

        **Automatic Actions:**
        1. Add queue member to database
        2. Deploy updated configuration to Asterisk
        3. Reload queue configuration in real-time

        **Note:** Configuration deployment happens automatically. No additional
        API calls are needed to make the queue member active.

        **Example curl:**
        ```bash
        curl -X POST ${YOUR_SERVER_URL}/api/v1/queues/QUEUE_UUID/members \
          -H "Authorization: Bearer YOUR_ORG_TOKEN" \
          -H "Content-Type: application/json" \
          -d '{
            "user_id": "USER_UUID",
            "penalty": 0
          }'
        ```
      parameters:
        - name: queueId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Queue UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id:
                  type: string
                  format: uuid
                  description: User UUID to add as queue member
                penalty:
                  type: integer
                  default: 0
                  description: Queue member penalty (0 = highest priority)
            example:
              user_id: "bf25246a-9434-4ef1-963f-2ac0fe6dcf9a"
              penalty: 0
      responses:
        '201':
          description: Member added successfully and configuration deployed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  queue_member:
                    $ref: '#/components/schemas/QueueMember'
                  configuration_deployed:
                    type: boolean
                    description: Indicates if Asterisk configuration was automatically deployed
              example:
                success: true
                message: "Queue member added successfully"
                queue_member:
                  id: "uuid-generated"
                  queue_id: "queue-uuid"
                  user_id: "bf25246a-9434-4ef1-963f-2ac0fe6dcf9a"
                  penalty: 0
                  paused: false
                configuration_deployed: true

    delete:
      security:
        - BearerAuth: []
      tags: [Queues]
      summary: Remove member from queue
      description: |
        Remove a user from a queue

        **Example curl:**
        ```bash
        curl -X DELETE "${YOUR_SERVER_URL}/api/v1/queues/QUEUE_UUID/members?userId=USER_UUID" \
          -H "Authorization: Bearer YOUR_ORG_TOKEN"
        ```
      parameters:
        - name: queueId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: userId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Member removed
...